requirements.txt
boto3>=1.28.0
pyyaml>=6.0
dnspython>=2.4.0
psycopg2-binary>=2.9.0
